requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
defusedxml==0.7.1
Django==3.2.7
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
httpie==2.5.0
idna==3.2
phonenumbers==8.12.31
Pillow==8.3.2
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.6
